requirements.txt
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
regex>=2023.0
numpy>=1.26.0
scikit-learn>=1.4.0
sentence-transformers>=2.5.0
torch>=2.2.0
openai>=1.40.0
pdf2image
pytesseract
PyPDF2
flask>=3.0.0
